requirements.txt
flask
gitpython
python-dotenv
anthropic
